Exercise C++ compiler within `libfoo` as well
